requirements.txt
# Core ML and NLP
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
datasets>=2.12.0
accelerate>=0.20.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
nltk>=3.8.0
spacy>=3.6.0

# Web Interface
streamlit>=1.25.0
streamlit-chat>=0.1.1
plotly>=5.15.0

# Evaluation Metrics
rouge-score>=0.1.2
sacrebleu>=2.3.0
bert-score>=0.3.13

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
pyyaml>=6.0
requests>=2.31.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
jupyter>=1.0.0

